Split out ser and de modules
